# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keasy API",
    "description": "Keasy — data pipeline execution and cloud storage API",
    "contact": {
      "name": "Ángel Iglesias Préstamo",
      "email": "angel.iglesias.prestamo@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/connections": {
      "get": {
        "tags": [
          "Connections"
        ],
        "operationId": "list_connections",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by connection type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connection"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connections"
        ],
        "operationId": "create_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Connection created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid connection or container not found"
          }
        }
      }
    },
...
    "/v1/connections/urls": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Sign GET URLs for locators, each with the credentials of the connection it\nlies under. Serves the editor and a job's run alike: a member reads what\nevery source connection holds, and the sink only through the job that\nwrote it. Public HTTP locators come back as they are.",
        "operationId": "sign_locators",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignLocatorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed GET URLs for the locators the caller may read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignLocatorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{id}": {
      "get": {
        "tags": [
          "Connections"
        ],
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Connections"
        ],
        "operationId": "delete_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Connection deleted"
          }
        }
      }
    },
    "/v1/connections/{id}/files": {
      "get": {
        "tags": [
          "Connections"
        ],
        "operationId": "list_connection_files",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of files in the connection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "File listing not supported"
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Connection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "location_type",
          "url"
        ],
        "properties": {
          "cloud_account_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "direction": {
            "$ref": "#/components/schemas/Direction",
            "description": "Read source vs the workspace write sink. Defaults to `source`."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConnectionKind"
          },
          "location_type": {
            "$ref": "#/components/schemas/LocationType"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ConnectionKind": {
        "type": "string",
        "enum": [
          "data",
          "vocab"
        ]
      },
...
      "CreateConnectionRequest": {
        "type": "object",
        "required": [
          "name",
          "kind",
          "location_type",
          "url"
        ],
        "properties": {
          "cloud_account_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "direction": {
            "$ref": "#/components/schemas/Direction",
            "description": "`source` (default) or `sink` (the owner output store; one per workspace)."
          },
          "kind": {
            "$ref": "#/components/schemas/ConnectionKind"
          },
          "location_type": {
            "$ref": "#/components/schemas/LocationType"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "Direction": {
        "type": "string",
        "description": "Whether a connection is a READ source (programs reference it via `@conn`) or\nthe workspace's WRITE sink (where the owner's job output is materialised).\nOrthogonal to [`ConnectionKind`] (which describes a source's data) and\n[`LocationType`]: a connection is a named, credentialed storage location, and\n`direction` says how it is used. Exactly one `sink` exists per workspace (the\nowner output store); `kind` is source-only and ignored for a sink.",
        "enum": [
          "source",
          "sink"
        ]
      },
...
      "FileEntry": {
        "type": "object",
        "required": [
          "path",
          "size"
        ],
        "properties": {
          "last_modified": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "LocationType": {
        "type": "string",
        "enum": [
          "cloud",
          "local"
        ]
      },
...
      "SignLocatorsRequest": {
        "type": "object",
        "required": [
          "locators"
        ],
        "properties": {
          "locators": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Locators fossil expanded from `@name/path` (`s3://bucket/prefix/users.csv`)."
          }
        }
      },
      "SignLocatorsResponse": {
        "type": "object",
        "required": [
          "urls"
        ],
        "properties": {
          "urls": {
            "type": "object",
            "description": "Locator → fetchable URL. A locator keasy will not sign is absent.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}